requirements.txt
flask==2.3.2
flask-sqlalchemy==3.0.3
flask-migrate==4.0.4
feedparser==6.0.10
beautifulsoup4==4.12.2
requests==2.31.0
python-dotenv==1.0.0
psycopg2-binary==2.9.6
sqlalchemy==2.0.15
